container/requirements.txt
flask==3.0.3
pyyaml==6.0.2